# flasky/static/openapi.yaml
openapi: 3.0.0
info:
  version: v3
  title: Store Manager
  description: >-
    Keep track of product stock by monitoring and managing sales using simple
    but efficient functions provided by the store manager API.
  contact:
    name: Store manager API
    url: 'https://that-store-manager-app-v2.herokuapp.com'
    email: awesomeme155@gmail.com
servers:
  - url: 'https://that-store-manager-app-v2.herokuapp.com/api/v2'
    description: Production server with postgres database
paths:
...
  /sales:
    post:
      tags:
        - Sales
      summary: A store attendant can make a sale
      description: ' Pass the required sale parameters through a post request to the /sales path .'
      parameters:
        - name: name
          in: query
          description: >-
            **attendant name**. *Example: Michael*. This is a required parameter and should be a valid string.
          schema:
            type: string
        - name: products
          in: query
          description: >-
           **products**. *Example: [products]*. This is a list of products or product for sale.
          schema:
            type: string

      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                title: success
                type: string
                example: success

        '400':
          description: Bad request
          content:
            application/json:
              schema:
                title: Content must be of type json
                type: string
                example: Content must be of type json

        '401':
          description: Unathourized
          content:
            application/json:
              schema:
                title: Attendant previllages required
                type: string
                example: Attendant previllages required
  /sales/:
    get:
      tags:
        - Sales
      summary: A store admin can fetch all sales
      description: ' Pass the required token parameter through a get request to the /sales path.'

      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                title: success
                type: string
                example: All sales successfully retrieved

        '400':
          description: Bad request
          content:
            application/json:
              schema:
                title: content type must be of type json
                type: string
                example: content type must be of type json

        '401':
          description: Unathourized
          content:
            application/json:
              schema:
                title: Admin previllages required
                type: string
                example: Admin previllages required
...
  /sales/attendant:
    get:
      tags:
        - Sales
      summary: A store attendant can fetch his/her sale record by id
      description: ' Pass the required id parameter through a get request to the /sales/attendant path .'
      parameters:
        - name: id
          in: query
          description: >-
            **sale id**. *Example: 12564356*. This is a required parameter and should be a valid number.
          schema:
            type: number
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                title: success
                type: string
                example: sale with ID 12564356 successfully retrieved

        '400':
          description: Bad request
          content:
            application/json:
              schema:
                title: content type must be of type json
                type: string
                example: content type must be of type json


        '401':
          description: Unathourized
          content:
            application/json:
              schema:
                title: Attendant previllages required
                type: string
                example: Attendant previllages required